// cypress/support/interfaces/user.interface.ts
export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
}

// Type for creating a new user (without id)
export type CreateUserRequest = Omit<User, 'id'>;

// Type for updating user (all fields are optional)
export type UpdateUserRequest = Partial<User>;